restructured parser and fix 12am wrong interpretation bug
