+++ src/main/java/timetable/TimeTableStorage.java
@@ -6,7 +6,6 @@ import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.FileWriter;
 import java.io.IOException;
-import java.text.ParseException;
 import java.time.format.DateTimeParseException;
 import java.util.Scanner;
